use yaml file; fix tests
