+++ hw03/src/main/java/ru/otus/spring/config/AppConfig.java
@@ -1,30 +1,9 @@
 package ru.otus.spring.config;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
 
-@PropertySource("classpath:application.properties")
-@Component
-public class AppConfig implements TestFileNameProvider, TestConfig {
-
-    private final String testFileName;
-
-    private final int rightAnswersCountToPass;
-
-    public AppConfig(@Value("${test.testFileName}") String testFileName,
-                     @Value("${test.rightAnswersCountToPass}") int rightAnswersCountToPass) {
-        this.testFileName = testFileName;
-        this.rightAnswersCountToPass = rightAnswersCountToPass;
-    }
-    
-    @Override
-    public String getTestFileName() {
-        return testFileName;
-    }
-
-    @Override
-    public int getRightAnswersCountToPass() {
-        return rightAnswersCountToPass;
-    }
+@EnableConfigurationProperties(TestProperties.class)
+@Configuration
+public class AppConfig {
 }
