+++ hw03/src/main/java/ru/otus/spring/config/TestProperties.java
@@ -1,0 +1,29 @@
+package ru.otus.spring.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.bind.ConstructorBinding;
+
+@ConfigurationProperties(prefix = "test")
+public class TestProperties implements TestFileNameProvider, TestConfig {
+
+    private final String testFileName;
+
+    private final int rightAnswersCountToPass;
+
+    @ConstructorBinding
+    public TestProperties(String testFileName,
+                          int rightAnswersCountToPass) {
+        this.testFileName = testFileName;
+        this.rightAnswersCountToPass = rightAnswersCountToPass;
+    }
+    
+    @Override
+    public String getTestFileName() {
+        return testFileName;
+    }
+
+    @Override
+    public int getRightAnswersCountToPass() {
+        return rightAnswersCountToPass;
+    }
+}
