--- hw03/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-test.testFileName=test.csv
-test.rightAnswersCountToPass=3
