+++ hw03/src/main/resources/application.yml
@@ -1,0 +1,3 @@
+test:
+  testFileName: test.csv
+  rightAnswersCountToPass: 3
