+++ hw03/src/test/resources/application.yml
@@ -1,0 +1,3 @@
+test:
+  commandLineRunner:
+    enabled: false
